Version 1.0
This script does a good number of things to reduce the amount of time spent setting up a computer for a new user inside of an organization. You can customize this script any way you like, and I've taken good care to document as much as I can where possible.